pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pachong"
version = "0.1.0"
description = "Dynamic website crawler MVP built with Playwright"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "markdownify>=1.1.0",
  "playwright>=1.53.0",
  "trafilatura>=2.0.0",
  "pydantic>=2.11.0",
  "PyYAML>=6.0.2",
  "selectolax>=0.3.26",
  "typer>=0.16.0",
]

[project.scripts]
pachong = "pachong.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
